fix(captcha): keep server env out of client bundles; dash captcha env on Railway
`@chia/api/captcha/constants` carries `X_CAPTCHA_RESPONSE` and `ErrorCode` so
browser code no longer imports the module that loads the server env. The dash
login form reads the captcha client env from its owner instead of the app env,
whose `adminEnv` extension is a server-key proxy on the client and threw
"Attempted to access a server-side environment variable" on first client use.

The captcha env resolves the provider once so the development test keys match
it. Railway IaC declares dash's `NEXT_PUBLIC_CAPTCHA_PROVIDER` as a reference to
`service` and preserves the site key; `.env.example` documents both.
