+++ apps/dash/.env.example
@@ -29,3 +29,9 @@ S3_ACCESS_KEY_ID=
 S3_SECRET_ACCESS_KEY=
 S3_REGION=auto
 S3_BUCKET_NAME=
+
+# Captcha on the login form; provider and site key must match `service`.
+NEXT_PUBLIC_CAPTCHA_PROVIDER="google-recaptcha"
+NEXT_PUBLIC_CAPTCHA_SITE_KEY=6LeIxAcTAAAAAJcZVRqyHh71UMIEGNQ_MXjiZKhI
+# NEXT_PUBLIC_CAPTCHA_PROVIDER="cloudflare-turnstile"
+# NEXT_PUBLIC_CAPTCHA_SITE_KEY=1x00000000000000000000AA
