+++ apps/www/src/components/agent/account-menu.tsx
@@ -8,7 +8,7 @@ import { useTranslations } from "next-intl";
 import { toast } from "sonner";
 
 import { agentQueryKeys } from "@chia/agent-elements/queries";
-import { X_CAPTCHA_RESPONSE } from "@chia/api/captcha";
+import { X_CAPTCHA_RESPONSE } from "@chia/api/captcha/constants";
 import { authClient } from "@chia/auth/client";
 
 import { SiteCaptcha } from "@/components/commons/captcha";
